refactor: move MultimodalSpecialTokens to constants.py, promote API detector to staticmethod
- Move MultimodalSpecialTokens from vl/model/base.py to vl/constants.py
  alongside Modality; fixes circular import and enables type annotations
  on mm_tokens params in preprocess_utils.py
- Promote _is_new_preprocess_api to MultimodalProcessor.@staticmethod,
  encapsulating the vl_encoder None guard inside the method
