+++ lmdeploy/serve/processors/multimodal.py
@@ -17,13 +17,6 @@ from lmdeploy.vl.media.video import VideoMediaIO
 logger = get_logger('lmdeploy')
 
 
-def _is_new_preprocess_api(model) -> bool:
-    """New-style preprocess takes `input_prompt` + `mm_processor_kwargs`,
-    legacy takes only `messages`."""
-    sig = inspect.signature(model.preprocess).parameters
-    return 'input_prompt' in sig and 'mm_processor_kwargs' in sig
-
-
 class MultimodalProcessor:
     """Processor for handling prompt preprocessing, message content merging,
     and multimodal processing."""
@@ -45,7 +38,16 @@ class MultimodalProcessor:
         self.chat_template = chat_template
         self.vl_encoder = vl_encoder
         self.backend = backend
-        self._uses_new_preprocess = _is_new_preprocess_api(vl_encoder.model) if vl_encoder else False
+        self._uses_new_preprocess = self._is_new_preprocess_api(vl_encoder)
+
+    @staticmethod
+    def _is_new_preprocess_api(vl_encoder) -> bool:
+        """New-style preprocess takes `input_prompt` + `mm_processor_kwargs`,
+        legacy takes only `messages`."""
+        if vl_encoder is None:
+            return False
+        sig = inspect.signature(vl_encoder.model.preprocess).parameters
+        return 'input_prompt' in sig and 'mm_processor_kwargs' in sig
 
     @staticmethod
     def merge_message_content(msg: dict) -> dict:
