+++ lmdeploy/vl/model/base.py
@@ -1,5 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-import dataclasses
 from abc import ABC, abstractmethod
 from itertools import groupby
 from typing import Any
@@ -447,25 +446,3 @@ class VisionModel(ABC):
         if arch and (arch == cls._arch or arch in cls._arch):
             return True
         return False
-
-
-@dataclasses.dataclass
-class MultimodalSpecialTokens:
-    image_token: str | list[str] | None = None
-    video_token: str | list[str] | None = None
-    audio_token: str | list[str] | None = None
-    ts_token: str | list[str] | None = None
-
-    image_token_id: int | None = None
-    video_token_id: int | None = None
-    audio_token_id: int | None = None
-    ts_token_id: int | None = None
-
-    def get_token_id_by_modality(self, modality: Modality) -> int | None:
-        """Get token ID for a given modality."""
-        return {
-            Modality.IMAGE: self.image_token_id,
-            Modality.VIDEO: self.video_token_id,
-            Modality.AUDIO: self.audio_token_id,
-            Modality.TIME_SERIES: self.ts_token_id,
-        }.get(modality)
